Refactor: 기존 네비게이션 타입 정의를 삭제하고 새로운 네비게이션 타입 및 Props 정의 추가
