--- src/types/navigation-props.ts
@@ -1,15 +1,0 @@
-import { NavigationDataType } from "@/types/navigation-data";
-
-export interface NavigationProps {
-  menuData: NavigationDataType;
-}
-
-export interface FullMenuProps {
-  sortedMenuData: NavigationDataType;
-  isMenuOpen: boolean;
-}
-
-export interface HamburgerButtonProps {
-  isMenuOpen: boolean;
-  onToggle: () => void;
-}
